% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetPlot.R
\name{adj_plot}
\alias{adj_plot}
\title{Plot network adjacency matrix}
\usage{
adj_plot(graph)
}
\arguments{
\item{graph}{an igraph object}
}
\description{
Plot network adjacency matrix
}
\examples{
graph <- netgen()
adj_plot(graph)

}
