% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asCharacterNetwork.R
\name{as.character.Network}
\alias{as.character.Network}
\title{Get network information as a character string.}
\usage{
\method{as.character}{Network}(x, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{character(1)}]
}
\description{
Get network information as a character string.
}

