% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateClusteredInstance.R
\name{as.data.frame.ClusteredNetwork}
\alias{as.data.frame.ClusteredNetwork}
\title{Convert cluster instance to data frame.}
\usage{
\method{as.data.frame}{ClusteredNetwork}(x, row.names = NULL,
  optional = FALSE, include.membership = TRUE, ...)
}
\arguments{
\item{x}{[\code{ClusterInstance}]\cr
Cluster instance.}

\item{row.names}{[\code{character}]\cr
Row names for the result. Default is \code{NULL}.}

\item{optional}{[any]\cr
Currently not used.}

\item{include.membership}{[\code{logical(1)}]\cr
Include the membership as a seperate column? Default is \code{TRUE}.}

\item{...}{[any]\cr
Currently not used.}
}
\description{
Convert cluster instance to data frame.
}

