% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asDataFrameNetwork.R
\name{as.data.frame.Network}
\alias{as.data.frame.Network}
\title{Convert (clustered) instance to data frame.}
\usage{
\method{as.data.frame}{Network}(x, row.names = NULL, optional = FALSE,
  include.extras = TRUE, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Cluster instance.}

\item{row.names}{[\code{character}]\cr
Row names for the result. Default is \code{NULL}.}

\item{optional}{[any]\cr
Currently not used.}

\item{include.extras}{[\code{logical(1)}]\cr
Include additional information like membership, types as seperate columns?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Currently not used.}
}
\description{
Convert (clustered) instance to data frame.
}

