% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateClusteredInstance.R
\name{autoplot.ClusterInstance}
\alias{autoplot.ClusterInstance}
\title{Autoplot function.}
\usage{
\method{autoplot}{ClusterInstance}(object, ...)
}
\arguments{
\item{object}{[\code{ClusterInstance}]\cr
Instance to visualize.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
  ggplot2 object.
}
\description{
Generates a \code{\link[ggplot2]{ggplot}} object. Nice possibility to
visualize 2-dimensional (clustered) data sets in the euclidean plane.
}

