% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplotNetwork.R
\name{autoplot.Network}
\alias{autoplot.Network}
\title{Autoplot function.}
\usage{
\method{autoplot}{Network}(object, ...)
}
\arguments{
\item{object}{[\code{Network}]\cr
Instance to visualize.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Generates a \code{\link[ggplot2]{ggplot}} object. Nice possibility to
visualize 2-dimensional (clustered) networks in the euclidean plane.
}

