% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitFive}
\alias{circuitFive}
\title{Item Family Five (Circuit)}
\usage{
circuitFive(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family five.
}
\details{
This function creates a plot with the network blueprint into your working directory.
Different even number nodes (Radical 2). 1 node with 4 edges (Sample Item 2). Minimunm is 5 nodes.
}
\examples{
circuitFive(1)

}
\author{
Aiden Loe
}
