% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitFour}
\alias{circuitFour}
\title{Item Family Four (Circuit)}
\usage{
circuitFour(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family four
}
\details{
This function creates a plot with the network blueprint into your working directory.
}
\examples{
circuitFour(1)

}
\author{
Aiden Loe
}
