% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitNine}
\alias{circuitNine}
\title{Item Family Nine (Circuit)}
\usage{
circuitNine(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family eight
}
\details{
This function creates a plot with the network blueprint into your working directory.
Different even number nodes (Radical 2). 1 node with 6 edges, 1 node with 4 edges, the rest with 2 edges.
}
\examples{
circuitNine(1)
}
\author{
Aiden Loe
}
