% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitOne}
\alias{circuitOne}
\title{Item Family One (Circuit)}
\usage{
circuitOne(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family one (Sample item 1)
}
\details{
This function creates a plot with the network blueprint into your working directory.
}
\examples{
circuitOne(1)

}
\author{
Aiden Loe
}
