% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitSeven}
\alias{circuitSeven}
\title{Item Family Seven (Circuit)}
\usage{
circuitSeven(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family seven
}
\details{
This function creates a plot with the network blueprint into your working directory.
Different even number nodes (Radical 2). 3 nodes with 4 edges
}
\examples{
circuitSeven(1)

}
\author{
Aiden Loe
}
