% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitSix}
\alias{circuitSix}
\title{Item Family Six (Circuit)}
\usage{
circuitSix(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family six
}
\details{
This function creates a plot with the network blueprint into your working directory.
Different even number nodes (Radical 2). 2 nodes with 4 edges
}
\examples{
circuitSix(1)

}
\author{
Aiden Loe
}
