% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitThree}
\alias{circuitThree}
\title{Item Family Three (Circuit)}
\usage{
circuitThree(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family three
}
\details{
This function creates a plot with the network blueprint into your working directory.
}
\examples{
circuitThree(1)

}
\author{
Aiden Loe
}
