% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_network.R
\name{circuitTwo}
\alias{circuitTwo}
\title{Item Family Two (Circuit)}
\usage{
circuitTwo(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family two
}
\details{
This function creates a plot with the network blueprint into your working directory.
}
\examples{
circuitTwo(1)

}
\author{
Aiden Loe
}
