% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_colours.R
\name{colour_display}
\alias{colour_display}
\title{Colour Display}
\usage{
colour_display(base.colour, start.colour, end.colour)
}
\arguments{
\item{base.colour}{This is the base colour selection.}

\item{start.colour}{This is the start colour selection.}

\item{end.colour}{This is the end colour selection.}
}
\description{
Selecting the colour display of the nodes
}
\details{
Selecting the colour display of the nodes. Again, not sure if this is necessary anymore since it does not affect the colour of the html, only the png image will be affected.
}
\author{
Aiden Loe
}
