% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportToCSVFormat.R
\name{exportToCSVFormat}
\alias{exportToCSVFormat}
\title{Exports a network to a simple comma separated file (CSV).}
\usage{
exportToCSVFormat(x, filename, full = TRUE, col.names = TRUE, digits = 2L)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network to export.}

\item{filename}{[\code{character(1)}]\cr
File name.}

\item{full}{[\code{logical(1)}]\cr
Should all meta data like cluster membership be stored. You most probably
want to use the default setting, which is \code{TRUE}.}

\item{col.names}{[\code{logical(1)}]\cr
Should there be a header row? Default is \code{TRUE}. Note that with this
option set to \code{FALSE}, it is not possible to import the file again.}

\item{digits}{[\code{integer(1)}]\cr
Round coordinates to this number of digits. Default is 2.}
}
\value{
Nothing
}
\description{
Exports a network to a simple comma separated file (CSV).
}
\note{
Currently we only support euclidean 2D instances.
}

