% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToFile.R
\name{exportToFile}
\alias{exportToFile}
\title{Exports a network to an proprietary format.}
\usage{
exportToFile(x, filename, digits = 2L)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network to export.}

\item{filename}{[\code{character(1)}]\cr
File name.}

\item{digits}{[\code{integer(1)}]\cr
Round coordinates to this number of digits. Default is 2.}
}
\value{
Nothing
}
\description{
The format used is similar to the TSPlib format (see \code{\link{exportToTSPlibFormat}}),
but it saves not only the point coordinates. It also saves the arrival times
of dynamic customers.
}

