% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterInstances.R
\name{filterInstancesBy}
\alias{filterInstancesBy}
\title{Filter instances.}
\usage{
filterInstancesBy(path, by = "DIMENSION", expr = NULL, full.names = FALSE)
}
\arguments{
\item{path}{[character(1)]
Path to instances in TSPlib format. All files with the file extension *.tsp
will be considered.}

\item{by}{[character(1)]
Keyword used for filtering, e.g., DIMENSION or EDGE_WEIGHT_TYPE.}

\item{expr}{[language(1)]
This must be a quoted expression with the \dQuote{value} variable used. The latter
is substiduted by the value of the \code{by} keyword before evaluation.}

\item{full.names}{[logical(1)]
  If \code{TRUE}, the directory path is prepended to the file names to give a
  relative file path. Otherwise, simply the file names are returned. Default is \code{FALSE}.}
}
\value{
[character]
  File names of the filtered files.
}
\description{
Offers the possibility to filter instances according to different specification
  parts of the TSPlib format, e.g., by dimension or by edge weight type.
}
\note{
This function does not expect a list of \code{Network} objects, but a
  path to a folder which contains TSP instances in the TSPlib format.
}
\examples{
\dontrun{
  source.dir = "."
  filterInstancesBy(source.dir, expr = quote(value <= 1000))
  filterInstancesByDimension(source.dir, expr = quote(value >= 50 && value <= 200))
  filterInstancesByEdgeWeightType(source.dir, expr = quote(!(value \%in\% "EUC_2D")))
  }
}
\seealso{
Other filter_functions: \code{\link{filterInstancesByDimension}};
  \code{\link{filterInstancesByEdgeWeightType}}
}

