% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInstances.R
\name{filterInstancesByEdgeWeightType}
\alias{filterInstancesByEdgeWeightType}
\title{Filter instances by edge weight type.}
\usage{
filterInstancesByEdgeWeightType(path, expr = NULL, full.names = TRUE)
}
\arguments{
\item{path}{[character(1)]
Path to instances in TSPlib format. All files with the file extension *.tsp
will be considered.}

\item{expr}{[language(1)]
This must be a quoted expression with the \dQuote{value} variable used. The latter
is substiduted by the value of the \code{by} keyword before evaluation.}

\item{full.names}{[logical(1)]
If \code{TRUE}, the directory path is prepended to the file names to give a
relative file path. Otherwise, simply the file names are returned. Default is \code{FALSE}.}
}
\seealso{
Other filter_functions: \code{\link{filterInstancesByDimension}},
  \code{\link{filterInstancesBy}}
}

