% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateGridNetwork.R
\name{generateGridNetwork}
\alias{generateGridNetwork}
\title{Generates a grid network in a hypercube.}
\usage{
generateGridNetwork(n.points = NULL, n.points.per.dim = NULL, n.dim = 2L,
  lower = 0, upper = 100)
}
\arguments{
\item{n.points}{[\code{integer(1)}]\cr
Number of points. Ignored if \code{n.points.per.dim} is not \code{NULL}.
Since the method generates only full grids, the root to base \code{n.dim}
must be an integer \code{integer}. We recommend to use the parameter
\code{n.points.per.dim} instead.}

\item{n.points.per.dim}{[\code{integer(1)}]\cr
Number of points in each dimension.}

\item{n.dim}{[\code{integer(1)}]\cr
Number of dimensions. Default ist 2.}

\item{lower}{[\code{numeric(1)}]\cr
Lower box constraint of cube. Default is \code{0}.}

\item{upper}{[\code{numeric(1)}]\cr
Upper box constraint of cube. Default is \code{100}.}
}
\value{
[\code{Network}]
}
\description{
Generates a grid network in a hypercube.
}
\examples{
x = generateGridNetwork(n.points.per.dim = 10L, upper = 50)
}

