% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateRandomInstance.R
\name{generateRandomInstance}
\alias{generateRandomInstance}
\title{Generates a random graph in a hypercube.}
\usage{
generateRandomInstance(n.points, n.dim = 2L, lower = 0, upper = 1)
}
\arguments{
\item{n.points}{[\code{integer(1)}]\cr
Number of points.}

\item{n.dim}{[\code{integer(1)}]\cr
Number of dimensions. Default ist 2.}

\item{lower}{[\code{numeric(1)}]\cr
Lower bound of cube.}

\item{upper}{[\code{numeric(1)}]\cr
Upper bound of cube.}
}
\value{
[\code{data.frame}]
}
\description{
Generates a random graph in a hypercube.
}

