% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDepotCoordinates.R
\name{getDepotCoordinates}
\alias{getDepotCoordinates}
\title{Get coordinates of depots.}
\usage{
getDepotCoordinates(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{matrix}]
}
\description{
Get coordinates of depots.
}
