% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumberOfClusters.R
\name{getNumberOfClusters}
\alias{getNumberOfClusters}
\title{Return the number of clusters of a network.}
\usage{
getNumberOfClusters(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{integer(1)}]
  Return the number of clusters.
}
\description{
Return the number of clusters of a network.
}
\note{
For simple random instances this function always returns 1.
}

