% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfDepots.R
\name{getNumberOfDepots}
\alias{getNumberOfDepots}
\title{Returns the number of depots of a network.}
\usage{
getNumberOfDepots(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{integer(1)}]
}
\description{
Returns the number of depots of a network.
}
