% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumberOfNodes.R
\name{getNumberOfNodes}
\alias{getNumberOfNodes}
\title{Returns number of nodes of a network.}
\usage{
getNumberOfNodes(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{integer(1)}]
  Number of nodes of the network.
}
\description{
Returns number of nodes of a network.
}

