% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getOptimalPointMatching.R
\name{getOptimalPointMatching}
\alias{getOptimalPointMatching}
\title{Computes optimal point assignment for two sets of points of equal size.}
\usage{
getOptimalPointMatching(coords1, coords2)
}
\arguments{
\item{coords1}{[\code{matrix}]\cr
Matrix of coordinates of the first point set.}

\item{coords2}{[\code{matrix}]\cr
Matrix of coordinates of the second point set.}
}
\value{
[\code{matrix}]
  Each row consists of the indizes of the pairwise matchings.
}
\description{
Internally it handles the points and the possible matchings as a bi-partite
graphs and finds an optimal matching due to euclidean distance by an
efficient linear programming solver.
}
\seealso{
\code{\link{visualizePointMatching}}
}

