% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineNumberOfPointsPerCluster.r
\name{getPointDistributionStrategies}
\alias{getPointDistributionStrategies}
\title{Returns the available strategies for distributing points around clusters.}
\usage{
getPointDistributionStrategies()
}
\value{
[\code{character}]
}
\description{
Returns the available strategies for distributing points around clusters.
}

