% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSPInstancesOverview.R
\name{getTSPInstancesOverview}
\alias{getTSPInstancesOverview}
\title{Get an overview of instances in a directory.}
\usage{
getTSPInstancesOverview(directory, append.filename = FALSE)
}
\arguments{
\item{directory}{[\code{character(1)}]\cr
Readable directory path.}

\item{append.filename}{[\code{logical(1)}]\cr
Should the full file names be appended to the data frame?
Default is \code{FALSE}.}
}
\value{
[\code{data.frame}]
}
\description{
This function expects a directory and returns a
data frame containing the most important properties, e. g., dimension,
edge weight type, of all TSPlib instances (with file extensions tsp) in
that directory. Moreover, the data frame contains information on the
availiability of the optimal tour length (files tsp.opt) and optimal
tour (tsp.tour).
}

