% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasDepots.R
\name{hasDepots}
\alias{hasDepots}
\title{Check if network has depots.}
\usage{
hasDepots(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if network has depots.
}

