% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importFromFile.R
\name{importFromFile}
\alias{importFromFile}
\title{Import a network from proprietary format.}
\usage{
importFromFile(filename)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
File name.}
}
\value{
Nothing
}
\description{
Import a network from proprietary format.
}

