% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEuclidean.R
\name{isEuclidean}
\alias{isEuclidean}
\title{Check if network is euclidean.}
\usage{
isEuclidean(x)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if a \code{Network} object has euclidean coordinates.
}
