% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{isNetwork}
\alias{isNetwork}
\title{Check if object is \code{Network}.}
\usage{
isNetwork(x)
}
\arguments{
\item{x}{[any]\cr
Arbitrary R object.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if object is \code{Network}.
}
