% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/morphInstances.R
\name{morphInstances}
\alias{morphInstances}
\title{Morphing of two networks with a convex combination of the coordinates.}
\usage{
morphInstances(x, y, alpha)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network or ClusteredNetwork.}

\item{y}{[\code{Network}]\cr
Network or ClusteredNetwork.}

\item{alpha}{[\code{numeric(1)}]\cr
Coeffiecient alpha for convex combination.}
}
\value{
[\code{Network}]
  Morphed Network instance.
}
\description{
This function takes two (clustered) networks with equal number of nodes,
and generates another instance by applying a convex combination to the coordinates
of node pairs. The node pairs are determined by a point matching algorithm,
which solves this assignement problem via a integer programming procedure.
If both instances contain depots, point matching is done seperately on depots
and the remaining nodes.
}

