% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netGen.R
\docType{package}
\name{netGen}
\alias{netGen}
\alias{netGen-package}
\title{netGen: A package for generating network type maze}
\description{
The mazeGen package provides functions to generate the Network Mazes. \cr
The item families are not designed to be increasing difficulty. They are just different at this stage. \cr
This is still at its early stages.
}
\references{
coming soon
}
