% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSampler.R
\name{NetSampler}
\alias{NetSampler}
\title{Network Sampling Routine}
\usage{
NetSampler(net_name = "samplenetwork.txt", out_name = "sampleout.txt",
  crit = c(1, 0), key_nodes = c(10, 50, 10, 1000), anfn = 0.5,
  numb_hidden = 0, hidden_modules = c(1, 5, 6, 0, 0, 0, 0, 0, 0, 0))
}
\arguments{
\item{net_name}{a name for the network}

\item{out_name}{a name for the output file}

\item{crit}{sampling criteria for key nodes and neighbors, see details}

\item{key_nodes}{number of key nodes to sample, from mi to mf at steps of delta-m
and number of realizations nr mi, mf, delta-m, nr}

\item{anfn}{number of first neighbors or fraction of first neighbors to include, see details}

\item{numb_hidden}{number of modules to exclude}

\item{hidden_modules}{list of modules to exclude (max 10 modules; only the first numb_hidden are used)}
}
\value{
output files are:
\itemize{
\item out_name -- main output file with info on the sunetwork
\item abund.txt / degree.txt / module.txt
\item subnet.txt  -- sampled network
\item netnodes.txt -- nodes in red or blue if belong subnet or not
\item netlinks.txt --  links in red or blue if connect subnet or not
}
}
\description{
Network Sampling Routine
}
\details{
afn argument can be written in numerous ways:
\itemize{
\item if 0 < anfn <= 1 it is interpreted as a fraction of the total number of neighbors
\item otherwise as the number of neighbors
\item to add all neighbors use 1.0
\item to add 1 neighbor per node use 1.1
\item to add 2 neighbours use 2, etc
}

sampling criteria for key nodes and neighbors

criterion for key nodes
\itemize{
\item 0 for random
\item 1 for lognormal
\item 2 for Fisher log series
\item 3 for exponential
\item 4 for degree
\item 5 for module
}

criterion for neighbors
\itemize{
\item 0 = random
\item 1 = weighted according to exponential distribution
}
}
