% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plot.logic.map}
\alias{plot.logic.map}
\title{plot.logic.map}
\usage{
\method{plot}{logic.map}(x, png = FALSE, v.size = 10, height = 1000,
  width = 1000, seed = NULL, png_name = "map.png",
  vertex.shape = "square", vertex.label.cex = 0.5,
  layout = layout_with_dh, xlab = ".", ylab = ".", cex.lab = 0.1)
}
\arguments{
\item{x}{This is the logic of the object we are goig to plot which is based on \code{closedMaps}.}

\item{png}{Tells you to print to your working directory automatically or not.}

\item{v.size}{Is the size of the vertice}

\item{height}{This is the height of the image}

\item{width}{This is the width of the image}

\item{seed}{This will always result in the same graph}

\item{png_name}{This is the input of the name}

\item{vertex.shape}{This determines the shape of the vertex shape}

\item{vertex.label.cex}{This is the size of the vertex label}

\item{layout}{Layout of the graph which graph}

\item{xlab}{The label of the x coordinates}

\item{ylab}{The label of the y coordinates}

\item{cex.lab}{The size of the labels}
}
\description{
To make it easier to plot the map in a png format.
}
\details{
Creating the the plot for the logic map.

Most of the internal functions are based on the igraph packge. This acts as a wrapper to the functions to make it easier for people to redesign the plots. However, the changes made here do not truly affect the way it is generated in html.
}
\author{
Aiden Loe

Aiden Loe
}
