% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.}
\alias{.}
\alias{quoted}
\title{Quote variables to create a list of unevaluated expressions for later evaluation.}
\usage{
.(...)
}
\arguments{
\item{...}{[any]
Unevaluated expressions to be recorded.}
}
\value{
List of symbol and language primitives.
}
\description{
This function is used by \code{filterTSPInstances} to pass
unevaluated expressions.
}

