% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rescaleNetwork.R
\name{rescaleNetwork}
\alias{rescaleNetwork}
\title{Normalize coordinates to unit cube maintaining its geography.}
\usage{
rescaleNetwork(x, method = "global2")
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{method}{[\code{character(1)}]\cr
Rescaling method which actually modifies the coordinates. ...}
}
\value{
[\code{Network}]
}
\description{
Normalize coordinates to unit cube maintaining its geography.
}

