% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailFive}
\alias{trailFive}
\title{Item Family Five (Trail)}
\usage{
trailFive(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family five (13 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailFive(1)


}
\author{
Aiden Loe
}
