% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailFour}
\alias{trailFour}
\title{Item Family Four (Trail)}
\usage{
trailFour(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family four (9 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailFour(1)


}
\author{
Aiden Loe
}
