% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailOne}
\alias{trailOne}
\title{Item Family One (Trail)}
\usage{
trailOne(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family one (6 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailOne(1)

}
\author{
Aiden Loe
}
