% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailSix}
\alias{trailSix}
\title{Item Family Six (Trail)}
\usage{
trailSix(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family six (10 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailSix(1)


}
\author{
Aiden Loe
}
