% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailThree}
\alias{trailThree}
\title{Item Family Three (Trail)}
\usage{
trailThree(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family three (10 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailThree(1)


}
\author{
Aiden Loe
}
