% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_network.R
\name{trailTwo}
\alias{trailTwo}
\title{Item Family Two (Trail)}
\usage{
trailTwo(value)
}
\arguments{
\item{value}{seed value}
}
\description{
This function generates the html template of the item family two (10 moves)
}
\details{
This function creates a plot with the network blueprint into your working directory. Same uneven number of nodes
}
\examples{
trailTwo(1)

}
\author{
Aiden Loe
}
