% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/morphInstances.R
\name{visualizeMorphing}
\alias{visualizeMorphing}
\title{Fancy visualization of morphing.}
\usage{
visualizeMorphing(x, y, alphas = c(0.25, 0.5, 0.75), arrows = TRUE)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network or ClusteredNetwork.}

\item{y}{[\code{Network}]\cr
Network or ClusteredNetwork.}

\item{alphas}{[\code{numeric}]\cr
Coeffiecients 'alpha' for different convex combinations.}

\item{arrows}{[\code{logical(1)}]\cr
Draw arrows originating in the points of \code{x} and ending in the
points matched in \code{y}. Default is \code{TRUE}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Takes two instances of equal size and some alphas. Computes the point matching
and morphings and visualizes the points of the first instance gradually move
towards their matched counterparts of the second instance.
}
\seealso{
\code{\link{morphInstances}}
}

