% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCharacterNetwork.R
\name{as.character.Network}
\alias{as.character.Network}
\title{Get basic network information as a string.}
\usage{
\method{as.character}{Network}(x, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{character(1)}]
}
\description{
Get basic network information as a string.
}
