% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDataFrameNetwork.R
\name{as.data.frame.Network}
\alias{as.data.frame.Network}
\title{Convert network to data frame.}
\usage{
\method{as.data.frame}{Network}(
  x,
  row.names = NULL,
  optional = FALSE,
  include.extras = TRUE,
  ...
)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{row.names}{[\code{character}]\cr
Row names for the result. Default is \code{NULL}.}

\item{optional}{[any]\cr
Currently not used.}

\item{include.extras}{[\code{logical(1)}]\cr
Include additional information like cluster membership and node type as specific columns?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{data.frame}]
}
\description{
Convert network to data frame.
}
\note{
If the instance contains of \eqn{n} depots, the depot coordinates fill the
  first \eqn{n} rows of the data frame.
}
