% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMatrixNetwork.R
\name{as.matrix.Network}
\alias{as.matrix.Network}
\title{Convert network to matrix.}
\usage{
\method{as.matrix}{Network}(x, ...)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}]
}
\description{
Convert network to matrix.
}
\note{
If the instance contains of \eqn{n} depots, the depot coordinates fill the
  first \eqn{n} rows of the matrix.
}

