% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportToFile.R
\name{exportToFile}
\alias{exportToFile}
\title{Exports a network to an proprietary format.}
\usage{
exportToFile(x, filename, digits = 2L)
}
\arguments{
\item{x}{[\code{Network}]\cr
Network to export.}

\item{filename}{[\code{character(1)}]\cr
File name.}

\item{digits}{[\code{integer(1)}]\cr
Round coordinates to this number of digits. Default is 2.}
}
\value{
Nothing
}
\description{
Exports a network to an proprietary format.
}

