% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateClusteredInstance.R
\name{generateClusteredInstance}
\alias{generateClusteredInstance}
\title{Function for generation of clustered instances.}
\usage{
generateClusteredInstance(n.cluster, n.points, n.dim = 2L,
  scale.factor = 1L, generator = lhs::maximinLHS, ...)
}
\arguments{
\item{n.cluster}{[\code{integer(1)}]\cr
Desired number of clusters.}

\item{n.points}{[\code{integer(1)}]\cr
Number of points for the instance.}

\item{n.dim}{[\code{integer(1)}]\cr
Number of dimensions. Most often you want to generate 2-dimensional instances
in the euclidean plane. Thus 2 is the default setting.}

\item{scale.factor}{[\code{numeric(1)}]\cr
Instances are generaded in the \code{n.dims}-dimensional
[0,1] hypercube. The numeric scaling parameter can be used to stretch dimensions
to arbitrary dimensions.}

\item{generator}{[\code{function}]\cr
Function which generates cluster centers. Default is \code{\link[lhs]{maximinLHS}}.}

\item{...}{[\code{any}]\cr
Not used yet.}
}
\value{
[\code{ClusterInstance}]
  Object of type \code{ClusterInstance}.
}
\description{
Testing algorithms for the Vehicle-Routing-Problem and understanding their
behaviour requires investigation of differnet instances. This function
generate clustered instances. It first generates n cluster centeres via a
LHS method and then distributes points to the clusters according to
gaussian distributions.
}

