% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGridNetwork.R
\name{generateGridNetwork}
\alias{generateGridNetwork}
\title{Generates a grid network.}
\usage{
generateGridNetwork(
  n.points.per.dim = NULL,
  n.dim = 2L,
  lower = 0,
  upper = 100,
  name = NULL
)
}
\arguments{
\item{n.points.per.dim}{[\code{integer(1)}]\cr
Number of points in each dimension.}

\item{n.dim}{[\code{integer(1)}]\cr
Number of dimensions. Default ist 2.}

\item{lower}{[\code{numeric(1)}]\cr
Lower box constaint for cube. Default is \code{0}.}

\item{upper}{[\code{numeric(1)}]\cr
Upper box constaint for cube. Default is \code{100}.}

\item{name}{[\code{character(1)} | \code{NULL}]\cr
Optional name for the generated network. Default is \code{NULL}. In this case
a random name is generated.}
}
\value{
[\code{Network}]
}
\description{
Generates a grid network.
}
\note{
Grid networks with depots are not supported at the moment.
}
\examples{
x = generateGridNetwork(n.points.per.dim = 10L, upper = 50)
}
