% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importFromTSPlibFormat.R
\name{importFromTSPlibFormat}
\alias{importFromTSPlibFormat}
\title{Import network from (extended) TSPlib format.}
\usage{
importFromTSPlibFormat(filename)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
Path to TSPlib file.}
}
\value{
[\code{Network}]
  Network object.
}
\description{
Import network from (extended) TSPlib format.
}
\note{
The extended TSPlib contains additional specification parts and a cluster
membership section. Currently only the import of symmetric TSP instances is possible.
}

