% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromTSPlibFormat.R
\name{importFromTSPlibFormat}
\alias{importFromTSPlibFormat}
\title{Import network from (extended) TSPlib format.}
\usage{
importFromTSPlibFormat(filename, round.distances = TRUE, read.opt = TRUE)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
Path to TSPlib file.}

\item{round.distances}{[\code{logical(1)}]\cr
Should the distances of EUC_2D instances be rounded to the nearest integer value?
Default is \code{TRUE}.}

\item{read.opt}{[\code{logical(1)}]\cr
Should the optimal tour length (in file filename.opt) and the optimal
tour (in file filename.tour) be loaded if avialable?
Default is \code{TRUE}.}
}
\value{
[\code{Network}]
  Network object.
}
\description{
Import network from (extended) TSPlib format.
}
\note{
The extended TSPlib contains additional specification parts and a cluster
membership section. Currently only the import of symmetric TSP instances is possible.
}
