% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeClusteredNetwork}
\alias{makeClusteredNetwork}
\title{Generates clustered network instances based in coordinates
and membership vector.}
\usage{
makeClusteredNetwork(coordinates, membership, types = NULL, lower = 0,
  upper = 1)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of 2D coordinates.}

\item{membership}{[\code{numeric}]\cr
Vector of memberships.}

\item{types}{[\code{character}]\cr
Vector of types for the nodes. For example to differentiate between
depots and customers in the context of vehicle routing.}

\item{lower}{[\code{numeric(1)}]\cr
Lower bound of bounding box.}

\item{upper}{[\code{numeric(1)}]\cr
Upper bound of bounding box.}
}
\value{
[\code{ClusteredNetwork}]
}
\description{
Generates clustered network instances based in coordinates
and membership vector.
}

