% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeClusteredNetwork}
\alias{makeClusteredNetwork}
\title{Generates clustered network.}
\usage{
makeClusteredNetwork(coordinates, depot.coordinates = NULL, membership,
  lower = NULL, upper = NULL)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of 2D coordinates.}

\item{depot.coordinates}{[\code{matrix} | NULL]\cr
Numeric matrix of 2D coordinates of depots. Default is \code{NULL}, which
means no depots at all.}

\item{membership}{[\code{numeric}]\cr
Vector of memberships.}

\item{lower}{[\code{numeric(1)}]\cr
Lower box constraint of cube. Default is 0.}

\item{upper}{[\code{numeric(1)}]\cr
Upper box constraint of cube. Default is 100.}
}
\value{
[\code{ClusteredNetwork}]
}
\description{
Generates clustered network.
}

