% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeClusteredNetwort}
\alias{makeClusteredNetwort}
\title{Generates clustered network instances based in coordinates
and membership vector.}
\usage{
makeClusteredNetwort(coordinates, membership, types = NULL)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of 2D coordinates.}

\item{membership}{[\code{numeric}]\cr
Vector of memberships.}

\item{types}{[\code{character}]\cr
Vector of types for the nodes. For example to differentiate between
depots and customers in the context of vehicle routing.}
}
\value{
[\code{ClusteredNetwork}]
}
\description{
Generates clustered network instances based in coordinates
and membership vector.
}

