% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeNetwork}
\alias{makeNetwork}
\title{Generate Network instance based on coordinates.}
\usage{
makeNetwork(coordinates, depot.coordinates = NULL, lower = 0, upper = 1)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of 2D coordinates.}

\item{depot.coordinates}{[\code{matrix} | NULL]\cr
Numeric matrix of 2D coordinates of depots. Default is \code{NULL}, which
means no depots at all.}

\item{lower}{[\code{numeric(1)}]\cr
Lower bound of bounding box.}

\item{upper}{[\code{numeric(1)}]\cr
Upper bound of bounding box.}
}
\value{
[\code{Network}]
}
\description{
Generate Network instance based on coordinates.
}

