% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeNetwork}
\alias{makeNetwork}
\title{Generate Network instance based on coordinates.}
\usage{
makeNetwork(coordinates)
}
\arguments{
\item{coordinates}{[\code{matrix}]\cr
Numeric matrix of 2D coordinates.}
}
\value{
[\code{Network}]
}
\description{
Generate Network instance based on coordinates.
}

