% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netGen.R
\docType{package}
\name{netGen}
\alias{netGen}
\alias{netGen-package}
\title{netGen: A package for generating network type maze}
\description{
The mazeGen package provides functions to generate the Network Mazes.

This is still at its early stages.
}
\references{
coming soon
}
