% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netGen.R
\name{netGen}
\alias{netGen}
\title{netGen: A package for generating network type maze}
\description{
The mazeGen package provides functions to generate the Network Mazes. \cr

The item families are not designed to be increasing difficulty. They are just different at this stage.
This package is still at its early stages.

Nevertheless, you can start generating your own html network maze by using \code{\link{netHTML}} and \code{\link{netHTML2arrows}}.
}
\references{
coming soon
}
