% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGen.R
\name{netgen}
\alias{netgen}
\title{netgen}
\usage{
netgen(name = "netgen", n_modav = c(50, 5), cutoffs = c(3, 0),
  net_type = 3, net_degree = 4, net_rewire = c(0.1, 0), mod_probs = 0)
}
\arguments{
\item{name}{a user-specified name for this network}

\item{n_modav}{network size and average module size (integer vector, length 2)}

\item{cutoffs}{module and submodule minimum sizes (integer vector, length 2).
(submodules are used only for bipartite and tripartite networks)}

\item{net_type}{integer indicating type, see details}

\item{net_degree}{average degree of connection}

\item{net_rewire}{global and local  network rewiring probabilities}

\item{mod_probs}{module probabilities for types 1 to 51,
used for constructing mixed networks, net_type = 0}
}
\value{
an `igraph` object
}
\description{
netgen function
}
\details{
network type
0 = mixed
1 = random
2 = scalefree
3 = nested
41 = bi-partite nested
42 = bi-partite random
51 = tri-trophic bipartite nested-random
52 = tri-trophic bipartite nested-bipartite nested
}
