% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSampler.R
\name{NetSampler}
\alias{NetSampler}
\title{Network Sampling Routine}
\usage{
NetSampler(net_name = "samplenetwork.txt", out_name = "sampleout.txt",
  crit = c(1, 0), key_nodes = c(10, 50, 10, 1000), anfn = 0.5,
  numb_hidden = 0, hidden_modules = c(1, 5, 6, 0, 0, 0, 0, 0, 0, 0))
}
\arguments{
\item{net_name}{a name for the network}

\item{out_name}{a name for the output file}

\item{crit}{sampling criteria for key nodes and neighbors, see details}

\item{key_nodes}{number of key nodes to sample, from mi to mf at steps of delta-m
and number of realizations nr mi, mf, delta-m, nr}

\item{anfn}{number of first neighbors or fraction of first neighbors to include, see details}

\item{numb_hidden}{number of modules to exclude}

\item{hidden_modules}{list of modules to exclude (max 10 modules; only the first numb_hidden are used)}
}
\value{
output files are:
- out_name -- main output file with info on the sunetwork
- abund.txt / degree.txt / module.txt
- subnet.txt  -- sampled network
- netnodes.txt -- nodes in red or blue if belong subnet or not
- netlinks.txt --  links in red or blue if connect subnet or not
}
\description{
Network Sampling Routine
}
\details{
afn argument can be written in numerous ways:
- if 0 < anfn <= 1 it is interpreted as a fraction of the total number of neighbors
- otherwise as the number of neighbors
- to add all neighbors use 1.0
- to add 1 neighbor per node use 1.1
- to add 2 neighbours use 2, etc

sampling criteria for key nodes and neighbors

criterion for key nodes

- 0 for random
- 1 for lognormal
- 2 for Fisher log series
- 3 for exponential
- 4 for degree
- 5 for module

criterion for neighbors

- 0 = random
- 1 = weighted according to exponential distribution
}
