% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizePointMatching.R
\name{visualizePointMatching}
\alias{visualizePointMatching}
\title{Visualize point matching.}
\usage{
visualizePointMatching(x, y, point.matching, highlight.longest = 0L)
}
\arguments{
\item{x}{[\code{matrix}]\cr
First \code{Network} object.}

\item{y}{[\code{matrix}]\cr
Second \code{Network} object.}

\item{point.matching}{[\code{matrix}]\cr
Point matching received via \code{getOptimalPointMatching}.}

\item{highlight.longest}{[\code{integer(1)}]\cr
Number of longest distances which should be particularly highlighted.
Default is \code{0}.}
}
\value{
[ggplot]
  ggplot2 object.
}
\description{
Draw the points and lines between the matched points for visualization.
}

