% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizePointMatching.R
\name{visualizePointMatching}
\alias{visualizePointMatching}
\title{Visualize point matching.}
\usage{
visualizePointMatching(x, y, point.matching, highlight.longest = 0L)
}
\arguments{
\item{x}{[\code{Network}]\cr
First network.}

\item{y}{[\code{Network}]\cr
Second network.}

\item{point.matching}{[\code{matrix}]\cr
Point matching received via \code{getOptimalPointMatching} for example.}

\item{highlight.longest}{[\code{integer(1)}]\cr
Number of longest distances which should be particularly highlighted.
Default is \code{0}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Draw the points and lines between the matched points for visualization.
}
\examples{
x = generateRandomNetwork(n.points = 20L, upper = 100)
y = generateClusteredNetwork(n.points = 20L, n.cluster = 2L, upper = 100)
\dontrun{
pm = getOptimalPointMatching(x$coordinates, y$coordinates)
print(visualizePointMatching(x, y, pm, highlight.longest = 2L))
}
}
\seealso{
\code{\link{getOptimalPointMatching}}, \code{\link{morphInstances}},
  \code{\link{visualizeMorphing}}
}

